06  - make sum int instead of double; add result fields for period and percents

+++ src/main/java/money/calculator/MoneyCalculator/entity/Deposit.java
@@ -16,7 +16,7 @@ public class Deposit {
     private String name;
     private Sex sex;
     private int age;
-    private double sum;
+    private int sum;
     private boolean employed;
     private boolean married;
     private int salary;
@@ -29,7 +29,7 @@ public class Deposit {
     }
 
 
-    public Deposit(String name, Sex sex, int age, double sum, boolean employed, boolean married, int salary, int period, double percents, double totalSum) {
+    public Deposit(String name, Sex sex, int age, int sum, boolean employed, boolean married, int salary, int period, double percents, double totalSum) {
         this.name = name;
         this.sex = sex;
         this.age = age;
@@ -75,11 +75,11 @@ public class Deposit {
         this.age = age;
     }
 
-    public double getSum() {
+    public int getSum() {
         return sum;
     }
 
-    public void setSum(double sum) {
+    public void setSum(int sum) {
         this.sum = sum;
     }
